feat!: migrates internal ocm plugin to external
Updates ocm cmd to interact with the C2P plugin manager

BREAKING CHANGE: the ocm command is now support with the
plugin manager only
