+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2025-03-15T01:53:20Z",
+  "generated_at": "2025-03-17T22:40:21Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -110,7 +110,7 @@
         "hashed_secret": "21c50805b553b7a40e48394a5d77d442587ddee2",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 143,
+        "line_number": 188,
         "type": "Secret Keyword",
         "verified_result": null
       }
