+++ cmd/ocm-plugin/server/helper.go
@@ -14,13 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ocm
+package server
 
 import (
-	typeconfigpolicy "github.com/oscal-compass/compliance-to-policy-go/v2/pkg/types/configurationpolicy"
-	typepolicy "github.com/oscal-compass/compliance-to-policy-go/v2/pkg/types/policy"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	typepolr "sigs.k8s.io/wg-policy-prototypes/policy-report/pkg/api/wgpolicyk8s.io/v1beta1"
+
+	typeconfigpolicy "github.com/oscal-compass/compliance-to-policy-go/v2/pkg/types/configurationpolicy"
+	typepolicy "github.com/oscal-compass/compliance-to-policy-go/v2/pkg/types/policy"
+	"github.com/oscal-compass/compliance-to-policy-go/v2/policy"
 )
 
 // - pass: the policy requirements are met
@@ -47,19 +49,19 @@ const (
 	PolicyResultSkip PolicyResult = "skip"
 )
 
-func mapToPolicyResult(complianceState typepolicy.ComplianceState) typepolr.PolicyResult {
-	var result PolicyResult
+func mapToPolicyResult(complianceState typepolicy.ComplianceState) policy.Result {
+	var result policy.Result
 	switch complianceState {
 	case typepolicy.Compliant:
-		result = PolicyResultPass
+		result = policy.ResultPass
 	case typepolicy.NonCompliant:
-		result = PolicyResultFail
+		result = policy.ResultFail
 	case typepolicy.Pending:
-		result = PolicyResultError
+		result = policy.ResultError
 	default:
-		result = PolicyResultError
+		result = policy.ResultError
 	}
-	return typepolr.PolicyResult(result)
+	return result
 }
 
 // Severity : low, medium, high, or critical
