--- cmd/ocm/result2oscal/main.go
@@ -1,30 +1,0 @@
-/*
-Copyright 2023 IBM Corporation
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"os"
-
-	"github.com/oscal-compass/compliance-to-policy-go/v2/cmd/ocm/result2oscal/cmd"
-)
-
-func main() {
-	err := cmd.New().Execute()
-	if err != nil {
-		os.Exit(1)
-	}
-}
