+++ docs/ocm/c2p-config.yaml
@@ -1,16 +1,8 @@
-compliance:
-  name: Demo Compliance
-  catalog:
-    url: https://raw.githubusercontent.com/usnistgov/oscal-content/main/nist.gov/SP800-53/rev5/json/NIST_SP-800-53_rev5_catalog.json
-  profile:
-    url: ./docs/ocm/oscal/profile.json
-  componentDefinition:
-    url: ./docs/ocm/oscal/component-definition.json
-policyResources: 
-  url: ./pkg/testdata/ocm/policies # example policy resources
-clusterGroups:
-  - name: myclusters
-    matchLabels:
-      my-cluster: 'true'
-target:
-  namespace: c2p
+component-definition: ./pkg/testdata/ocm/component-definition.json
+plugins:
+  ocm:
+    policy-dir: ./pkg/testdata/ocm/policies
+    policy-results-dir: ./pkg/testdata/ocm/policy-results
+    temp-dir: /tmp/ocm
+    policy-set-name: "kubernetes"
+    namespace: "test"
