+++ docs/tools/README.md
@@ -1,0 +1,6 @@
+6. Prettify OSCAL Assessment Results in .md format
+    ```
+    c2pcli tools oscal2posture -c ./docs/ocm/plugin-example.yaml --assessment-results /tmp/assessment-results.json -o /tmp/compliance-posture.md
+    ```
+
+You can view an example compliance posture like [compliance-posture.md](../ocm/final-outputs/compliance-posture.md)
